requirements_dev.txt
pylint==2.1.1
pytest==3.8.0
pytest-flask==0.12.0
pytest-mock==1.10.0
yapf==0.24.0